upd: moved str transform methods to prevent cicular deps
