+++ src/models/cmyk/index.ts
@@ -1,3 +1,2 @@
 export { cmyk2rgb } from "./transform";
 export { parseCMYKColor, parseCMYKString } from "./parse";
-export { rgb2cmykString } from "./util";
