+++ src/models/cmyk/util.ts
@@ -1,6 +1,5 @@
 import { clamp, round } from "@util/helpers";
-import type { ColorCMYK, ColorRGB } from "../../types";
-import { rgb2cmyk } from "../rgb";
+import type { ColorCMYK } from "../../types";
 
 /**
  * Clamps the CMYK color object values.
@@ -26,14 +25,4 @@ export function roundCMYK({ c, m, y, k, a = 1 }: ColorCMYK): ColorCMYK {
 		k: round(k, 2),
 		a: round(a, 2)
 	};
-}
-
-/**
- * Converts RGB Color object to CMYK string.
- */
-export function rgb2cmykString(color: ColorRGB): string {
-	const { c, m, y, k, a } = roundCMYK(rgb2cmyk(color));
-	return a && a < 1
-		? `device-cmyk(${c}% ${m}% ${y}% ${k}% / ${a})`
-		: `device-cmyk(${c}% ${m}% ${y}% ${k}%)`;
 }
