+++ src/models/hsl/index.ts
@@ -1,3 +1,3 @@
 export { hsl2hsv, hsl2rgb } from "./transform";
 export { parseHSLColor, parseHSLString } from "./parse";
-export { roundHSL, rgb2hslString } from "./util";
+export { roundHSL } from "./util";
