+++ src/models/hsv/index.ts
@@ -1,3 +1,3 @@
 export { hsv2hsl, hsv2rgb } from "./transform";
 export { parseHSVColor } from "./parse";
-export { roundHSV, rgb2hsvString } from "./util";
+export { roundHSV } from "./util";
