+++ src/models/hsv/util.ts
@@ -1,6 +1,5 @@
 import { clamp, clampDegrees, round } from "@util/helpers";
-import type { ColorHSV, ColorRGB } from "../../types";
-import { rgb2hsv } from "../rgb";
+import type { ColorHSV } from "../../types";
 
 /**
  * Clamps the HSV color object values.
@@ -24,15 +23,4 @@ export function roundHSV({ h, s, v, a = 1 }: ColorHSV): ColorHSV {
 		v: round(v),
 		a: round(a)
 	};
-}
-
-/**
- * Transforms the RGB color object to HSL color string.
- * Functional whitespace syntax is used.
- */
-export function rgb2hsvString(rgb: ColorRGB): string {
-	const { h, s, v, a = 1 } = roundHSV(rgb2hsv(rgb));
-	return a < 1
-		? `hsv(${h}deg ${s}% ${v}% / ${a})`
-		: `hsv(${h}deg ${s}% ${v}%)`;
 }
