+++ src/models/xyz/index.ts
@@ -2,9 +2,5 @@ export { xyz2rgb } from "./transform";
 export { parseXYZColor } from "./parse";
 export {
 	clampXYZ,
-	roundXYZ,
-	adaptXYZtoD50,
-	adaptXYZtoD65,
-	D50,
-	rgb2xyzString
+	roundXYZ
 } from "./util";
